Logic for sending output CV in relation to input voltage is working

+++ CubeIDE/ModulinV2/Drivers/DACx0501/DACx0501.h
@@ -21,8 +21,8 @@ typedef struct {
 
     SPI_HandleTypeDef *spi;
 
-    uint8_t currentValue;
-    uint8_t newValue;
+    uint16_t currentValue;
+    uint16_t newValue;
 } dacx0501;
 
 // Function prototypes
